Compiler 2.0 43rd pass on rewrite - Got Virtual Machine to build, mass testing and bug fixes begin

+++ lib/runtime/jit/Jit.h
@@ -65,7 +65,7 @@ public:
 private:
     Thread *tSelf;
     recursive_mutex mutex;
-    _List<int64_t> messageQueue;
+    _List<int32_t> messageQueue;
     x64Assembler* x64Asm;
     int error;
 };
